# amaima/backend/openapi.yaml
openapi: 3.1.0
info:
  title: AMAIMA Public Simulation API
  version: 1.0.0
  description: |
    Public endpoint for inspecting AMAIMA's routing decisions in simulation mode.
    - No execution occurs.
    - Outputs are explanatory and watermarked as simulated/non-authoritative.
    - For real execution, use private /v1/query (auth-required, not public).

servers:
  - url: https://api.amaima.com/v1  # Prod; swap for staging/local

paths:
  /v1/query:
    post:
      summary: Execute Query Routing
      description: |
        Routes and executes a query. Requires API key authentication.
      operationId: executeRoute
      security:
        - ApiKeyAuth: []
      requestBody:
        $ref: '#/components/schemas/SimulateRequest'
      responses:
        '200':
          description: "Successful query execution"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '401':
          description: "Unauthorized"
        '500':
          description: "Internal Server Error"

  /v1/simulate:
    post:
      summary: Simulate Query Routing
      description: |
        Returns a simulated routing decision with explanations.
        - Watermarked as simulated: true.
        - No actual model inference or cost incurred.
        - Borderline cases flagged; confidence decomposed.
      operationId: simulateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateRequest'
      responses:
        '200':
          description: Simulated routing decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateResponse'
        '400':
          description: Invalid query (e.g., empty or malformed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited (public sim capped at 10/min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal simulation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SimulateRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The query text to simulate routing for.
          minLength: 1
          maxLength: 2000  # Tune for abuse prevention
          example: "Design a scalable microservices architecture..."

    SimulateResponse:
      type: object
      required:
        - simulated
        - complexity_level
        - model
        - execution_mode
        - confidence
        - reasons
      properties:
        simulated:
          type: boolean
          description: Always true for public simulation.
          enum: [true]
        execution:
          type: string
          description: No execution in simulation.
          enum: ["none"]
        confidence_scope:
          type: string
          description: Scope limited to explanatory.
          enum: ["explanatory"]
        query_hash:  # Privacy anchor
          type: string
          description: Anonymized hash of input query.
          example: "sha256:abc123..."
        complexity_level:
          type: string
          enum: ["TRIVIAL", "SIMPLE", "MODERATE", "COMPLEX", "EXPERT", "BORDERLINE_ADVANCED_EXPERT"]  # Your levels + bands
        model:
          type: string
          example: "gpt4-turbo"
        execution_mode:
          type: string
          enum: ["batch_parallel", "parallel_min_latency", "streaming_real_time"]  # Your refined modes
        security_level:
          type: string
          enum: ["low", "medium", "high", "paranoid"]
        latency_estimate_ms:
          type: integer
          example: 150
        cost_estimate_usd:
          type: number
          example: 0.015
        confidence:
          type: object
          properties:
            complexity:
              type: number
              example: 0.85
            model_fit:
              type: number
              example: 0.92
            execution_fit:
              type: number
              example: 0.88
            overall:
              type: number
              example: 0.88
        reasons:
          type: object
          properties:
            complexity_reason:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    enum: ["ABSTRACT_REASONING", "TECHNICAL_DEPTH", "MULTI_STEP_LOGIC"]
                  label:
                    type: string
                    example: "abstract reasoning detected"
            model_reason:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    enum: ["LATENCY_OPTIMAL", "COST_EFFICIENT"]
                  label:
                    type: string
            execution_reason:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    enum: ["PARALLEL_MIN_LATENCY", "STREAMING_REAL_TIME"]
                  label:
                    type: string

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Invalid query: must be non-empty string."

    ExecuteResponse:
      allOf:
        - $ref: '#/components/schemas/SimulateResponse'
        - type: object
          properties:
            simulated:
              enum: [false]
            actual_latency_ms:
              type: integer
              description: "Measured post-execution."
              example: 142
            actual_cost_usd:
              type: number
              description: "Incurred cost."
              example: 0.012
            output:
              type: string
              description: "Model-generated response."
              example: "Architecture design: ..."